+++ src/app/actions.ts
@@ -2,6 +2,7 @@
 
 import { validateName } from "@/lib/formValidation";
 import { createClient } from "@/utils/supabase/server";
+import { normalizeReferrer } from "@/utils/referrer";
 import { getBaseUrl } from "@/utils/url";
 import {
   encodedRedirect,
@@ -12,24 +13,6 @@ import { revalidatePath } from "next/cache";
 import { headers } from "next/headers";
 import { redirect } from "next/navigation";
 import { getTranslations } from "next-intl/server";
-
-const normaliseReferrer = (referrer: string | undefined) => {
-  if (!referrer) return undefined;
-
-  let current = referrer;
-
-  for (let i = 0; i < 3; i += 1) {
-    try {
-      const decoded = decodeURIComponent(current);
-      if (decoded === current) break;
-      current = decoded;
-    } catch {
-      break;
-    }
-  }
-
-  return current;
-};
 
 export const signUpAction = async (formData: FormData, request?: Request) => {
   const t = await getTranslations("Errors");
@@ -46,7 +29,7 @@ export const signUpAction = async (formData: FormData, request?: Request) => {
   const origin = headersList.get("origin");
 
   // Get attribution data
-  const referrer = normaliseReferrer(
+  const referrer = normalizeReferrer(
     formData.get("initial_referrer")?.toString()
   );
   const utmSource = formData.get("utm_source")?.toString();
